componentize modal
