+++ containers/tefca-viewer/src/app/query/components/header/header.module.css
@@ -1,0 +1,8 @@
+.dataUsagePolicyButton {
+  cursor: pointer;
+  background: none !important;
+  border: none !important;
+  color: white !important;
+  padding: 0px !important;
+  font: inherit !important;
+}
