+++ containers/tefca-viewer/src/app/query/components/resultsView/ResultsViewTable.tsx
@@ -34,6 +34,7 @@ const ResultsViewTable: React.FC<ResultsViewTable> = ({ accordionItems }) => {
                 }
                 expanded={true}
                 id={titleId}
+                key={titleId}
                 headingLevel={"h3"}
                 accordionClassName={styles.accordionWrapper}
                 containerClassName={styles.accordionContainer}
